Assets
Added the asset class.

+++ src/com/game/code/graphics/Screen.java
@@ -53,7 +53,7 @@ public class Screen extends Canvas {
 		g.fillRect(0, 0, getWidth(), getHeight());
 		g.setColor(Color.GREEN);
 		g.drawString("ups: " + GameLoop.frames, 0, 10);
-		g.drawString("Height: " + GameState.heightReached / 60, 0, 20);
+		g.drawString("Height: " + (int)(GameState.heightReached / 60), 0, 20);
 		RunnableObject.drawObjects(g); //Renders objects we create.
 		
 		g.dispose(); //Disposes of the graphics context to save memory space.
